Remove C function getXStringSetClass()
C function getXStringSetClass() was a wrapper around C function
get_XStringSet_xsbaseclassname() defined in the Biostrings package,
which is itself a wrapper around C function get_List_elementType()
defined in the S4Vectors package. However, get_XStringSet_xsbaseclassname()
is about to be removed from Biostrings so any call to it should be replaced
with a call to get_List_elementType(). But since getXStringSetClass() is
not used, I removed it instead of fixing it (the fix would have been to
make it a wrapper around get_List_elementType() instead of around
get_XStringSet_xsbaseclassname()).